delivery_service/requirements.txt
grpcio
protobuf
grpcio-tools
redis